Make Dragabor front match back better
The back sprite seems to have a longer neck. Also, one of the paws was missing claws - or, if intentional, the other paw has claws that don't match the other paw's lack. Also, the curve of the horn was hard to see, at least on my screen, making it look like the horn was shorter than the back sprite version.
This has the side benefit of making it not look like Dragonite crossed with Barney.