Starter Walking Sprites
Sideways Lambert base sprite are *not* by me. The other Lambert are however, so you can blame me for how bad they look - I was started to get fatigued at that point doing starters.
 I put them in a new folder because the old one was incredibly crowded with human sprites and it made it hard for me to overview all of them just by looking at the folder itself. Rockitten is apparently already in the old folder and referenced elsewhere so I didn't move it.
I tested out a really basic followscript using these guys. It has some issues with collisions with other characters trying to pathfind near to the player, so be sure to save often if playtesting this. Teleporting to the player can leave a collision up also.